requirements.txt
torch
torchvision
numpy
matplotlib
pillow
opencv-python
scikit-learn
tqdm
notebook
jupyter
onnx
onnxruntime
onnxsim
